+++ src/Game/Scenes.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+enum Scene{
+    Game,
+    Menu,
+    GameOver
+};
+
